requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.1
qrcode>=7.4.2
